src/app/user/components/user-detail/user-detail.component.ts
import { Component, OnInit } from '@angular/core';
import { UserService } from '../../services/user.service';
import { ActivatedRoute, Router } from '@angular/router';

@Component({
  selector: 'app-user-detail',
  templateUrl: './user-detail.component.html',
  styleUrls: ['./user-detail.component.scss'],
})
export class UserDetailComponent implements OnInit {
  userInfo?: IUser;

  constructor(private userService: UserService, private route: ActivatedRoute, private router: Router) {}

  ngOnInit(): void {
    this.route.params.subscribe((params) => {
      this.userService.get(params['id']!).then((user) => {
        this.userInfo = user;
      });
    });
  }

  onBack() {
    this.router.navigate(['/user']);
  }
}
